Splitting files to types, arranged by area.
