+++ SgmlReaderDll/EntityContent/DesktopEntityResolver.cs
@@ -6,90 +6,6 @@ using System.Text;
 
 namespace Sgml
 {
-    class FileEntityContent : IEntityContent
-    {
-        readonly string path;
-
-        public FileEntityContent(string path)
-        {
-            this.path = path;
-        }
-
-        public Encoding Encoding => Encoding.UTF8;
-
-        public string MimeType => string.Empty;
-
-        public Uri Redirect => new Uri(path);
-
-        public Stream Open()
-        {
-            return new FileStream(path, FileMode.Open, FileAccess.Read);
-        }
-    }
-
-    class WebEntityContent : IEntityContent
-    {
-        readonly WebResponse response;
-
-        public WebEntityContent(WebResponse response)
-        {
-            this.response = response;
-        }
-
-        public Encoding Encoding
-        {
-            get
-            {
-                string contentType = response.ContentType.ToLowerInvariant();
-                int i = contentType.IndexOf("charset");
-                Encoding e = Encoding.UTF8;
-                if (i >= 0)
-                {
-                    int j = contentType.IndexOf("=", i);
-                    int k = contentType.IndexOf(";", j);
-                    if (k < 0)
-                        k = contentType.Length;
-
-                    if (j > 0)
-                    {
-                        j++;
-                        string charset = contentType.Substring(j, k - j).Trim();
-                        try
-                        {
-                            e = Encoding.GetEncoding(charset);
-                        }
-                        catch (ArgumentException)
-                        {
-                        }
-                    }
-                }
-                return e;
-            }
-        }
-
-        public string MimeType
-        {
-            get
-            {
-                string contentType = response.ContentType.ToLowerInvariant();
-                string mimeType = contentType;
-                int i = contentType.IndexOf(';');
-                if (i >= 0)
-                {
-                    mimeType = contentType.Substring(0, i);
-                }
-                return mimeType;
-            }
-        }
-
-        public Uri Redirect => response.ResponseUri;
-
-        public Stream Open()
-        {
-            return response.GetResponseStream();
-        }
-    }
-
     /// <summary>
     /// A special implementation for IEntityResolver that uses System.IO to access local files
     /// and HttpWebRequest for web content.
