+++ SgmlReaderDll/EntityContent/EmbeddedResourceEntityContent.cs
