+++ SgmlReaderDll/Enums/AttributePresence.cs
@@ -1,0 +1,44 @@
+﻿/*
+ * Copyright (c) 2020 Microsoft Corporation. All rights reserved.
+ * Modified work Copyright (c) 2008 MindTouch. All rights reserved. 
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+#if WINDOWS_DESKTOP
+using System.Runtime.Serialization;
+using System.Security.Permissions;
+#endif
+
+namespace Sgml
+{
+    /// <summary>
+    /// Defines the different constraints on an attribute's presence on an element.
+    /// </summary>
+    public enum AttributePresence
+    {
+        /// <summary>
+        /// The attribute has a default value, and its presence is optional.
+        /// </summary>
+        Default,
+
+        /// <summary>
+        /// The attribute has a fixed value, if present.
+        /// </summary>
+        Fixed,
+
+        /// <summary>
+        /// The attribute must always be present on every element.
+        /// </summary>
+        Required,
+        
+        /// <summary>
+        /// The element is optional.
+        /// </summary>
+        Implied
+    }
+}
