+++ SgmlReaderDll/Enums/DeclaredContent.cs
@@ -1,0 +1,44 @@
+﻿/*
+ * Copyright (c) 2020 Microsoft Corporation. All rights reserved.
+ * Modified work Copyright (c) 2008 MindTouch. All rights reserved. 
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+#if WINDOWS_DESKTOP
+using System.Runtime.Serialization;
+using System.Security.Permissions;
+#endif
+
+namespace Sgml
+{
+    /// <summary>
+    /// Where nested subelements cannot occur within an element, its contents can be declared to consist of one of the types of declared content contained in this enumeration.
+    /// </summary>
+    public enum DeclaredContent
+    {
+        /// <summary>
+        /// Not defined.
+        /// </summary>
+        Default,
+        
+        /// <summary>
+        /// Character data (CDATA), which contains only valid SGML characters.
+        /// </summary>
+        CDATA,
+        
+        /// <summary>
+        /// Replaceable character data (RCDATA), which can contain text, character references and/or general entity references that resolve to character data.
+        /// </summary>
+        RCDATA,
+        
+        /// <summary>
+        /// Empty element (EMPTY), i.e. having no contents, or contents that can be generated by the program.
+        /// </summary>
+        EMPTY
+    }
+}
