+++ SgmlReaderDll/Enums/GroupType.cs
@@ -1,0 +1,44 @@
+﻿/*
+ * Copyright (c) 2020 Microsoft Corporation. All rights reserved.
+ * Modified work Copyright (c) 2008 MindTouch. All rights reserved. 
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+#if WINDOWS_DESKTOP
+using System.Runtime.Serialization;
+using System.Security.Permissions;
+#endif
+
+namespace Sgml
+{
+    /// <summary>
+    /// The type of the content model group, defining the order in which child elements can occur.
+    /// </summary>
+    public enum GroupType
+    {
+        /// <summary>
+        /// No model group.
+        /// </summary>
+        None,
+        
+        /// <summary>
+        /// All elements must occur, in any order.
+        /// </summary>
+        And,
+        
+        /// <summary>
+        /// One (and only one) must occur.
+        /// </summary>
+        Or,
+        
+        /// <summary>
+        /// All element must occur, in the specified order.
+        /// </summary>
+        Sequence 
+    };
+}
