+++ SgmlReaderDll/Enums/LiteralType.cs
@@ -1,0 +1,39 @@
+﻿/*
+ * Copyright (c) 2020 Microsoft Corporation. All rights reserved.
+ * Modified work Copyright (c) 2008 MindTouch. All rights reserved. 
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+#if WINDOWS_DESKTOP
+using System.Runtime.Serialization;
+using System.Security.Permissions;
+#endif
+
+namespace Sgml
+{
+    /// <summary>
+    /// The different types of literal text returned by the SgmlParser.
+    /// </summary>
+    public enum LiteralType
+    {
+        /// <summary>
+        /// CDATA text literals.
+        /// </summary>
+        CDATA,
+
+        /// <summary>
+        /// SDATA entities.
+        /// </summary>
+        SDATA,
+
+        /// <summary>
+        /// The contents of a Processing Instruction.
+        /// </summary>
+        PI
+    };
+}
