+++ SgmlReaderDll/SgmlParseException.cs
@@ -1,0 +1,112 @@
+/*
+ * Copyright (c) 2020 Microsoft Corporation. All rights reserved.
+ * Modified work Copyright (c) 2008 MindTouch. All rights reserved. 
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Globalization;
+using System.IO;
+#if WINDOWS_DESKTOP
+using System.Runtime.Serialization;
+using System.Security.Permissions;
+#endif
+using System.Text;
+using System.Xml;
+
+namespace Sgml
+{
+    /// <summary>
+    /// Thrown if any errors occur while parsing the source.
+    /// </summary>
+#if WINDOWS_DESKTOP
+    [Serializable]
+#endif
+    public class SgmlParseException : Exception
+    {
+        private readonly string m_entityContext;
+
+        /// <summary>
+        /// Instantiates a new instance of SgmlParseException with no specific error information.
+        /// </summary>
+        public SgmlParseException()
+        {
+        }
+
+        /// <summary>
+        /// Instantiates a new instance of SgmlParseException with an error message describing the problem.
+        /// </summary>
+        /// <param name="message">A message describing the error that occurred</param>
+        public SgmlParseException(string message)
+            : base(message)
+        {
+        }
+
+        /// <summary>
+        /// Instantiates a new instance of SgmlParseException with an error message describing the problem.
+        /// </summary>
+        /// <param name="message">A message describing the error that occurred</param>
+        /// <param name="e">The entity on which the error occurred.</param>
+        public SgmlParseException(string message, Entity e)
+            : base(message)
+        {
+            if (e != null)
+                m_entityContext = e.Context();
+        }
+
+        /// <summary>
+        /// Instantiates a new instance of SgmlParseException with an error message describing the problem.
+        /// </summary>
+        /// <param name="message">A message describing the error that occurred</param>
+        /// <param name="innerException">The original exception that caused the problem.</param>
+        public SgmlParseException(string message, Exception innerException)
+            : base(message, innerException)
+        {
+        }
+
+#if WINDOWS_DESKTOP
+        /// <summary>
+        /// Initializes a new instance of the SgmlParseException class with serialized data. 
+        /// </summary>
+        /// <param name="streamInfo">The object that holds the serialized object data.</param>
+        /// <param name="streamCtx">The contextual information about the source or destination.</param>
+        protected SgmlParseException(SerializationInfo streamInfo, StreamingContext streamCtx)
+            : base(streamInfo, streamCtx)
+        {
+            if (streamInfo != null)
+                m_entityContext = streamInfo.GetString("entityContext");
+        }
+#endif
+
+        /// <summary>
+        /// Contextual information detailing the entity on which the error occurred.
+        /// </summary>
+        public string EntityContext => m_entityContext;
+
+#if WINDOWS_DESKTOP
+        /// <summary>
+        /// Populates a SerializationInfo with the data needed to serialize the exception.
+        /// </summary>
+        /// <param name="info">The <see cref="SerializationInfo"/> to populate with data. </param>
+        /// <param name="context">The destination (see <see cref="StreamingContext"/>) for this serialization.</param>
+        [SecurityPermission(SecurityAction.Demand, SerializationFormatter=true)]
+        public override void GetObjectData(SerializationInfo info, StreamingContext context)
+        {
+            if (info is null)
+                throw new ArgumentNullException(nameof(info));
+
+            info.AddValue("entityContext", m_entityContext);
+            base.GetObjectData(info, context);
+        }
+#endif
+    }
+
+    
+}
