+++ SgmlReaderDll/SgmlReader.csproj
@@ -6,6 +6,7 @@
     <LangVersion>latest</LangVersion>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <Deterministic>True</Deterministic>
+    <RootNamespace>Sgml</RootNamespace>
   </PropertyGroup>
   
   <PropertyGroup Label="Signing">
