+++ SgmlReaderDll/SgmlReader/SgmlReader.cs
@@ -20,323 +20,6 @@ using System.Xml;
 
 namespace Sgml
 {
-    /// <summary>
-    /// SGML is case insensitive, so here you can choose between converting
-    /// to lower case or upper case tags.  "None" means that the case is left
-    /// alone, except that end tags will be folded to match the start tags.
-    /// </summary>
-    public enum CaseFolding
-    {
-        /// <summary>
-        /// Do not convert case, except for converting end tags to match start tags.
-        /// </summary>
-        None,
-
-        /// <summary>
-        /// Convert tags to upper case.
-        /// </summary>
-        ToUpper,
-
-        /// <summary>
-        /// Convert tags to lower case.
-        /// </summary>
-        ToLower
-    }
-
-    /// <summary>
-    /// Due to the nature of adjacent self-closing and auto-closing SGML elements the <see cref="SgmlReader"/>'s default behavior may cause unwanted whitespace to appear at either end of the element's text content. This enum instructs <see cref="SgmlReader"/> that such whitespace is not significant-whitespace.<br />
-    /// Note that this is a <see cref="FlagsAttribute"/>-enum.
-    /// </summary>
-    /// <remarks>Internally strings are not actually trimmed using <see cref="String.Trim(char[])"/>, but instead by skipping leading and trailing whitespace during the read-phase - this avoids unnecessary string copying and heap-allocation.</remarks>
-    [Flags]
-    public enum TextWhitespaceHandling
-    {
-        /// <summary>
-        /// <c>0x00</c> - All leading and trailing whitespace in <c>InnerText</c> and <c>Value</c> properties is kept verbatim. This is the default behavior.
-        /// </summary>
-        None = 0,
-        
-        /// <summary>
-        /// <c>0x01</c> - Leading whitespace will not be present in <c>TextContent</c> and <c>Value</c> properties.
-        /// </summary>
-        TrimLeading  = 1,
-
-        /// <summary>
-        /// <c>0x02</c> - Trailing whitespace will not be present in <c>TextContent</c> and <c>Value</c> properties.
-        /// </summary>
-        TrimTrailing = 2,
-
-        /// <summary>
-        /// <c>0x04</c> - This flag only takes effect when either or both <see cref="TrimLeading"/> and <see cref="TrimTrailing"/> are set. When this flag is set, only line-break characters <c>\r</c> (CR) and <c>\n</c> (LF) will be trimmed.
-        /// </summary>
-        OnlyLineBreaks = 4,
-
-        /// <summary>
-        /// <c>0x03</c> - (This is a combination of flags). Neither leading nor trailing whitespace will be be present in <c>TextContent</c> and <c>Value</c> properties.
-        /// </summary>
-        TrimBoth  = TrimLeading | TrimTrailing,
-
-        /// <summary>
-        /// <c>0x06</c> - (This is a combination of flags). Only trailing <c>\r</c> (CR) and <c>\n</c> (LF) characters will be trimmed from trailing whitespace in in <c>InnerText</c> and <c>Value</c> properties.
-        /// </summary>
-        OnlyTrailingLineBreaks = OnlyLineBreaks | TrimTrailing
-    }
-
-    internal static class EnumExtensions
-    {
-        // Bitwise operators on enums to check for flags are much faster than Enum.HasFlag(), unfortunately.
-
-        [System.Runtime.CompilerServices.MethodImpl(System.Runtime.CompilerServices.MethodImplOptions.AggressiveInlining)]
-        public static bool HasFlagBits(this TextWhitespaceHandling value, TextWhitespaceHandling bits) => (value & bits) == bits;
-    }
-
-    /// <summary>
-    /// This stack maintains a high water mark for allocated objects so the client
-    /// can reuse the objects in the stack to reduce memory allocations, this is
-    /// used to maintain current state of the parser for element stack, and attributes
-    /// in each element.
-    /// </summary>
-    internal sealed class HWStack<T>
-        where T: class
-    {
-        private T[] _items;
-        private int _size;
-        private int _count;
-        private readonly int _growth;
-
-        /// <summary>
-        /// Initialises a new instance of the HWStack class.
-        /// </summary>
-        /// <param name="growth">The amount to grow the stack space by, if more space is needed on the stack.</param>
-        public HWStack(int growth)
-        {
-            _growth = growth;
-        }
-
-        /// <summary>
-        /// The number of items currently in the stack.
-        /// </summary>
-        public int Count
-        {
-            get => _count;
-            set => _count = value;
-        }
-
-        /// <summary>
-        /// The size (capacity) of the stack.
-        /// </summary>
-        public int Size => _size;
-
-        /// <summary>
-        /// Returns the item at the requested index or null if index is out of bounds
-        /// </summary>
-        /// <param name="i">The index of the item to retrieve.</param>
-        /// <returns>The item at the requested index or null if index is out of bounds.</returns>
-        public T this[int i]
-        {
-            get => (i >= 0 && i < _size) ? _items[i] : null;
-            set => _items[i] = value;
-        }
-
-        /// <summary>
-        /// Removes and returns the item at the top of the stack
-        /// </summary>
-        /// <returns>The item at the top of the stack.</returns>
-        public T Pop()
-        {
-            _count--;
-            if (_count > 0)
-            {
-                return _items[_count - 1];
-            }
-
-            return default(T);
-        }
-
-        /// <summary>
-        /// Pushes a new slot at the top of the stack.
-        /// </summary>
-        /// <returns>The object at the top of the stack.</returns>
-        /// <remarks>
-        /// This method tries to reuse a slot, if it returns null then
-        /// the user has to call the other Push method.
-        /// </remarks>
-        public T Push()
-        {
-            if (_count == _size)
-            {
-                int newsize = _size + _growth;
-                T[] newarray = new T[newsize];
-                if (_items != null)
-                    Array.Copy(_items, newarray, _size);
-
-                _size = newsize;
-                _items = newarray;
-            }
-            return _items[_count++];
-        }
-
-        /// <summary>
-        /// Remove a specific item from the stack.
-        /// </summary>
-        /// <param name="i">The index of the item to remove.</param>
-        public void RemoveAt(int i)
-        {
-            _items[i] = default(T);
-            Array.Copy(_items, i + 1, _items, i, _count - i - 1);
-            _count--;
-        }
-    }
-
-    /// <summary>
-    /// This class represents an attribute.  The AttDef is assigned
-    /// from a validation process, and is used to provide default values.
-    /// </summary>
-    internal class Attribute
-    {
-        internal string Name;    // the atomized name.
-        internal AttDef DtdType; // the AttDef of the attribute from the SGML DTD.
-        internal char QuoteChar; // the quote character used for the attribute value.
-        private string _literalValue; // the attribute value
-
-        /// <summary>
-        /// Attribute objects are reused during parsing to reduce memory allocations, 
-        /// hence the Reset method.
-        /// </summary>
-        public void Reset(string name, string value, char quote)
-        {
-            Name = name;
-            _literalValue = value;
-            QuoteChar = quote;
-            DtdType = null;
-        }
-
-        public string Value
-        {
-            get
-            {
-                if (_literalValue != null) 
-                    return _literalValue;
-                if (DtdType != null) 
-                    return DtdType.Default;
-                return null;
-            }
-/*            set
-            {
-                this.m_literalValue = value;
-            }*/
-        }
-
-        public bool IsDefault => _literalValue is null;
-    }    
-
-    /// <summary>
-    /// This class models an XML node, an array of elements in scope is maintained while parsing
-    /// for validation purposes, and these Node objects are reused to reduce object allocation,
-    /// hence the reset method.  
-    /// </summary>
-    [DebuggerDisplay("Sgml.Node, Type: {NodeType}, Name: {Name}, Value: {Value}")]
-    internal sealed class Node
-    {
-        internal XmlNodeType NodeType;
-        internal string Value;
-        internal XmlSpace Space;
-        internal string XmlLang;
-        internal bool IsEmpty;        
-        internal string Name;
-        internal ElementDecl DtdType; // the DTD type found via validation
-        internal State CurrentState;
-        internal bool Simulated; // tag was injected into result stream.
-        private readonly HWStack<Attribute> _attributes = new (10);
-
-        /// <summary>
-        /// Attribute objects are reused during parsing to reduce memory allocations, 
-        /// hence the Reset method. 
-        /// </summary>
-        public void Reset(string name, XmlNodeType nt, string value) 
-        {           
-            Value = value;
-            Name = name;
-            NodeType = nt;
-            Space = XmlSpace.None;
-            XmlLang= null;
-            IsEmpty = true;
-            _attributes.Count = 0;
-            DtdType = null;
-        }
-
-        public Attribute AddAttribute(string name, string value, char quotechar, bool caseInsensitive) 
-        {
-            Attribute a;
-            // check for duplicates!
-            for (int i = 0, n = _attributes.Count; i < n; i++)
-            {
-                a = _attributes[i];
-                if (string.Equals(a.Name, name, caseInsensitive ? StringComparison.OrdinalIgnoreCase : StringComparison.Ordinal))
-                {
-                    return null;
-                }
-            }
-            // This code makes use of the high water mark for attribute objects,
-            // and reuses exisint Attribute objects to avoid memory allocation.
-            a = _attributes.Push();
-            if (a is null)
-            {
-                a = new Attribute();
-                _attributes[_attributes.Count-1] = a;
-            }
-            a.Reset(name, value, quotechar);
-            return a;
-        }
-
-        public void RemoveAttribute(string name)
-        {
-            for (int i = 0, n = _attributes.Count; i < n; i++)
-            {
-                Attribute a  = _attributes[i];
-                if (string.Equals(a.Name, name, StringComparison.OrdinalIgnoreCase))
-                {
-                    _attributes.RemoveAt(i);
-                    return;
-                }
-            }
-        }
-        public void CopyAttributes(Node n) 
-        {
-            for (int i = 0, len = n._attributes.Count; i < len; i++)
-            {
-                Attribute a = n._attributes[i];
-                Attribute na = this.AddAttribute(a.Name, a.Value, a.QuoteChar, false);
-                na.DtdType = a.DtdType;
-            }
-        }
-
-        public int AttributeCount => _attributes.Count;
-
-        public int GetAttribute(string name) 
-        {
-            for (int i = 0, n = _attributes.Count; i < n; i++) 
-            {
-                Attribute a = _attributes[i];
-                if (string.Equals(a.Name, name, StringComparison.OrdinalIgnoreCase)) 
-                {
-                    return i;
-                }
-            }
-            return -1;
-        }
-
-        public Attribute GetAttribute(int i) 
-        {
-            if (i>=0 && i<_attributes.Count)
-            {
-                Attribute a = _attributes[i];
-                return a;
-            }
-            return null;
-        }
-    }
-
     internal enum State
     {
         Initial,    // The initial state (Read has not been called yet)
@@ -474,7 +157,7 @@ namespace Sgml
                      // no need to hit the w3.org servers in this case
                      || _syslit is "http://www.w3.org/TR/html4/loose.dtd")
                 {
-                    if (_docType != null && StringUtilities.EqualsIgnoreCase(_docType, "html"))
+                    if (_docType != null && Extensions.EqualsIgnoreCase(_docType, "html"))
                     {
                         Stream stream = null;
                         // see if our resolver can find it.
@@ -504,7 +187,7 @@ namespace Sgml
                     CaseFolding.ToLower => dtdName.ToLowerInvariant(),
                     _ => dtdName
                 };
-                _isHtml = StringUtilities.EqualsIgnoreCase(dtdName, "html");
+                _isHtml = Extensions.EqualsIgnoreCase(dtdName, "html");
             }
         }
